♻️ Organize Everest related code into src/everest/*.rs
