+++ src/everest/installer.rs
@@ -6,8 +6,8 @@ use std::{
     process::{Command, Stdio},
 };
 
-/// Runs MiniInstaller.
-pub fn run(root_dir: &Path) -> io::Result<()> {
+/// Install Everest by running MiniInstaller.
+pub fn install(root_dir: &Path) -> io::Result<()> {
     let installer = MiniInstaller::new(root_dir);
     installer.grant_execute_permission()?;
     installer.execute()
